requirements.txt
os
secrets
python-dotenv
fastapi
uvicorn
sentence-transformers